pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "co-cli"
version = "0.4.5"
description = "Co — personal AI assistant CLI with approval-gated shell, Google Drive/Gmail/Calendar, Obsidian, and OTel tracing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasette>=0.65.2",
    "google-api-python-client>=2.189.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "google-genai>=1.61.0",
    "opentelemetry-sdk>=1.39.1",
    "prompt-toolkit>=3.0.52",
    "pydantic-ai==1.59.0",
    "rich>=14.3.2",
    "typer>=0.21.1",
    "datasette-pretty-json>=0.3",
    "httpx>=0.28.1",
    "html2text>=2025.4.15",
    "rapidfuzz>=3.14.3",
    "pysqlite3>=0.6.0",
    "sqlite-vec>=0.1.6",
    "ollama>=0.6.1",
]

[project.scripts]
co = "co_cli.main:app"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
reranker = [
    "fastembed>=0.4.0",
]

[tool.pytest.ini_options]
addopts = "-x --durations=0"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
